Move FetchRegionService to billing
