+++ app/services/billing/fetch-region.service.js
@@ -5,7 +5,7 @@
  * @module FetchRegionService
  */
 
-const RegionModel = require('../../../models/water/region.model.js')
+const RegionModel = require('../../models/water/region.model.js')
 
 /**
  * Fetches the region with the matching NALD Region ID
