+++ test/services/billing/fetch-region.service.test.js
@@ -8,11 +8,11 @@ const { describe, it, beforeEach } = exports.lab = Lab.script()
 const { expect } = Code
 
 // Test helpers
-const DatabaseHelper = require('../../../support/helpers/database.helper.js')
-const RegionHelper = require('../../../support/helpers/water/region.helper.js')
+const DatabaseHelper = require('../../support/helpers/database.helper.js')
+const RegionHelper = require('../../support/helpers/water/region.helper.js')
 
 // Thing under test
-const FetchRegionService = require('../../../../app/services/billing/supplementary/fetch-region.service.js')
+const FetchRegionService = require('../../../app/services/billing/fetch-region.service.js')
 
 describe('Fetch Region service', () => {
   const naldRegionId = 9
